api/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sib-api-v3-sdk
requests==2.31.0
bcrypt==4.1.2
